car_showroom/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-countries==7.5
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
typing-extensions==4.4.0
psycopg2-binary==2.8.5
